add permission checking

+++ src/main/java/com/iconectiv/irsf/portal/controller/ListServiceController.java
@@ -1,20 +1,26 @@
 package com.iconectiv.irsf.portal.controller;
 
 import com.iconectiv.irsf.portal.config.CustomerContextHolder;
+import com.iconectiv.irsf.portal.core.PermissionRole;
+import com.iconectiv.irsf.portal.model.common.UserDefinition;
 import com.iconectiv.irsf.portal.model.customer.ListDefintion;
 import com.iconectiv.irsf.portal.service.ListService;
 import com.iconectiv.irsf.util.JsonHelper;
+
+import java.util.Map;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-
 
 @Controller
 class ListServiceController extends BaseRestController {
@@ -23,14 +29,19 @@ class ListServiceController extends BaseRestController {
 	@Autowired
 	private ListService listService;
 
-	@RequestMapping(value = "/list/{schema}/{listName}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+	@RequestMapping(value = "/list/{listName}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public ResponseEntity<String> getListDetails(@PathVariable String schema, @PathVariable String listName) {
+	public ResponseEntity<String> getListDetails(@RequestHeader Map<String, String> header, @PathVariable String listName) {
 		ResponseEntity<String> rv;
 		try {
-			CustomerContextHolder.setSchema(schema);
+			UserDefinition loginUser = getLoginUser(header);
+			assertAuthorized(loginUser, PermissionRole.CustAdmin.value());
+
+			CustomerContextHolder.setSchema(loginUser.getSchemaName());
 			ListDefintion listDef = listService.getListDetails(listName);
-            rv = makeSuccessResult(listDef);
+			rv = makeSuccessResult(listDef);
+		} catch (SecurityException e) {
+			rv = makeErrorResult(e, HttpStatus.FORBIDDEN);
 		} catch (Exception e) {
 			rv = makeErrorResult(e);
 		}
@@ -41,23 +52,28 @@ class ListServiceController extends BaseRestController {
 		return rv;
 	}
 
-	//TO support junit testing
-    @RequestMapping(value = "/list/{schema}/{listName}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-    @ResponseBody
-    public ResponseEntity<String> deleteListByNameRequest(@PathVariable String schema, @PathVariable String listName) {
-        ResponseEntity<String> rv;
-        try {
-			CustomerContextHolder.setSchema(schema);
-            listService.deleteListDefinition(listName);
-            rv = makeSuccessResult();
-        } catch (Exception e) {
-            rv = makeErrorResult(e);
-        }
+	// TO support junit testing
+	@RequestMapping(value = "/list/{listName}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@ResponseBody
+	public ResponseEntity<String> deleteListByNameRequest(@RequestHeader Map<String, String> header, @PathVariable String listName) {
+		ResponseEntity<String> rv;
+		try {
+			UserDefinition loginUser = getLoginUser(header);
+			assertAuthorized(loginUser, PermissionRole.CustAdmin.value());
 
-        if (log.isDebugEnabled()) {
-            log.debug(JsonHelper.toJson(rv));
-        }
-        return rv;
-    }
+			CustomerContextHolder.setSchema(loginUser.getSchemaName());
+			listService.deleteListDefinition(listName);
+			rv = makeSuccessResult();
+		} catch (SecurityException e) {
+			rv = makeErrorResult(e, HttpStatus.FORBIDDEN);
+		} catch (Exception e) {
+			rv = makeErrorResult(e);
+		}
+
+		if (log.isDebugEnabled()) {
+			log.debug(JsonHelper.toJson(rv));
+		}
+		return rv;
+	}
 
 }
